legacy/flask_backend/requirements.txt
Flask>=1.0.3
Flask-Bootstrap>=3.3.7.1
Flask-Login>=0.4.1
Flask-Mail>=0.9.1
Flask-Migrate>=2.4.0
Flask-SQLAlchemy>=2.4.0
Flask-WTF>=0.14.2
pyjwt>=2.6.0
python-dotenv>=0.10.3
email-validator>=1.1.2